# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIError:
    properties:
      errorCode:
        type: integer
      errorMessage:
        type: string
    type: object
  schemas.Category:
    properties:
      categoryName:
        example: Category 1
        type: string
      description:
        example: Description 1
        type: string
    type: object
  schemas.CategoryResponse:
    properties:
      active:
        example: false
        type: boolean
      categoryName:
        example: Category 1
        type: string
      description:
        example: Description 1
        type: string
      id:
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        type: string
    type: object
  schemas.CreditCard:
    properties:
      brand:
        example: Visa
        type: string
      customerID:
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        type: string
      cvv:
        example: "847"
        type: string
      mm:
        example: "07"
        type: string
      number:
        example: "4468467578025601"
        type: string
      yyyy:
        example: "2025"
        type: string
    type: object
  schemas.CreditCardResponse:
    properties:
      brand:
        example: Visa
        type: string
      customerID:
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        type: string
      cvv:
        example: "847"
        type: string
      id:
        example: 5cc1f36f-1287-4c88-63fb-601feb9634be
        type: string
      mm:
        example: "07"
        type: string
      number:
        example: "4468467578025601"
        type: string
      yyyy:
        example: "2025"
        type: string
    type: object
  schemas.Customer:
    properties:
      address:
        example: 6597 Westheimer Rd
        type: string
      birthday:
        example: "1990-01-02"
        type: string
      city:
        example: Billings
        type: string
      country:
        example: United States
        type: string
      dni:
        example: "105883636"
        type: string
      email:
        example: ramona.ellis@example.com
        type: string
      name:
        example: Ramona Ellis
        type: string
      password:
        example: addison
        type: string
      phone:
        example: "7097900"
        type: string
      postalCode:
        example: "63104"
        type: string
      userName:
        example: yellowpeacock117
        type: string
    type: object
  schemas.CustomerResponse:
    properties:
      address:
        example: 6597 Westheimer Rd
        type: string
      birthday:
        example: "1990-01-02"
        type: string
      city:
        example: Billings
        type: string
      country:
        example: United States
        type: string
      dni:
        example: "105883636"
        type: string
      email:
        example: ramona.ellis@example.com
        type: string
      id:
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        type: string
      name:
        example: Ramona Ellis
        type: string
      password:
        example: addison
        type: string
      phone:
        example: "7097900"
        type: string
      postalCode:
        example: "63104"
        type: string
      userName:
        example: yellowpeacock117
        type: string
    type: object
...
host: go-rest-api.onrender.com
info:
  contact:
    email: mario.diaz@correounivalle.edu.co
    name: AlexVonEinzbern
    url: https://github.com/AlexVonEinzbern
  description: ecommerce api-rest with basic CRUD operations (Create, Read, Update
    and Delete) products, suppliers, etc. Manage automatically invoice and shipping
    operations
  license:
    name: Apache 2.0
  termsOfService: http://swagger.io/terms/
  title: go-rest-api
  version: "1"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: Search all catagories in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.CategoryResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all Categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a Category
      parameters:
      - description: Category type
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/schemas.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CategoryResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by id
      parameters:
      - description: Category ID
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete Category by id
      tags:
      - Categories
    patch:
      consumes:
      - application/json
      description: Update a category by id
      parameters:
      - description: Category ID
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/schemas.CategoryResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update Category by id
      tags:
      - Categories
  /categories/active:
    get:
      consumes:
      - application/json
      description: Search active categories in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.CategoryResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all active Categories
      tags:
      - Categories
  /creditcard:
    get:
      consumes:
      - application/json
      description: Search all credit cards in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.CreditCardResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all SearchCreditCards
      tags:
      - CreditCards
    post:
      consumes:
      - application/json
      description: Create a CreditCard
      parameters:
      - description: CreditCard a subcategory
        in: body
        name: creditcard
        required: true
        schema:
          $ref: '#/definitions/schemas.CreditCard'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CreditCardResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create creditcard
      tags:
      - CreditCards
  /creditcard/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a credit card by id
      parameters:
      - description: CreditCard ID
        example: 5cc1f36f-1287-4c88-63fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete CreditCard by id
      tags:
      - CreditCards
    get:
      consumes:
      - application/json
      description: Search CreditCard by id in the DataBase
      parameters:
      - description: CreditCard ID
        example: 5cc1f36f-1287-4c88-63fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.CreditCardResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search CreditCard by id
      tags:
      - CreditCards
    patch:
      consumes:
      - application/json
      description: Update a credit card by id
      parameters:
      - description: CreditCard ID
        example: 5cc1f36f-1287-4c88-63fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/schemas.CreditCardResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update CreditCard by id
      tags:
      - CreditCards
  /customers:
    get:
      consumes:
      - application/json
      description: Search all customers in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.CustomerResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all Customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Create a Customer
      parameters:
      - description: Customer type
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/schemas.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CustomerResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Customer
      tags:
      - Customers
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a customer by id
      parameters:
      - description: Customer ID
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete Customer by id
      tags:
      - Customers
    get:
      consumes:
      - application/json
      description: Search Customer by id in the DataBase
      parameters:
      - description: Customer ID
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.CustomerResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search Customer by id
      tags:
      - Customers
    patch:
      consumes:
      - application/json
      description: Update a customer by id
      parameters:
      - description: Customer ID
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/schemas.CustomerResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update Customer by id
      tags:
      - Customers
...
swagger: "2.0"